Conflicts:
	LayoutExpressions/Expression.swift
	LayoutExpressions/ItemDimensionArgument.swift
	LayoutExpressions/UIViewExtensions.swift

+++ LayoutExpressions/ItemDimensionArgument.swift
@@ -3,7 +3,7 @@
 import UIKit
 
 class ItemDimensionArgument: ItemAttributeArgument {
-	init(item: AnyObject, attribute: NSLayoutAttribute, multiplier: Float? = nil, constant: Float? = nil) {
+	init(item: AnyObject, attribute: NSLayoutAttribute, multiplier: CGFloat? = nil, constant: CGFloat? = nil) {
 		assert(attribute == .Width || attribute == .Height)
 		super.init(item: item, attribute: attribute, multiplier: multiplier, constant: constant)
 	}
@@ -11,17 +11,17 @@ class ItemDimensionArgument: ItemAttributeArgument {
 
 // MARK: Comparison Operators
 
-func ==(lhs: ItemDimensionArgument, constant: Float) -> Expression<ItemDimensionArgument, ConstantArgument> {
+func ==(lhs: ItemDimensionArgument, constant: CGFloat) -> Expression<ItemDimensionArgument, ConstantArgument> {
 	let rhs = ConstantArgument(constant: constant)
 	return Expression(lhs: lhs, relation: .Equal, rhs: rhs)
 }
 
-func <=(lhs: ItemDimensionArgument, constant: Float) -> Expression<ItemDimensionArgument, ConstantArgument> {
+func <=(lhs: ItemDimensionArgument, constant: CGFloat) -> Expression<ItemDimensionArgument, ConstantArgument> {
 	let rhs = ConstantArgument(constant: constant)
 	return Expression(lhs: lhs, relation: .LessThanOrEqual, rhs: rhs)
 }
 
-func >=(lhs: ItemDimensionArgument, constant: Float) -> Expression<ItemDimensionArgument, ConstantArgument> {
+func >=(lhs: ItemDimensionArgument, constant: CGFloat) -> Expression<ItemDimensionArgument, ConstantArgument> {
 	let rhs = ConstantArgument(constant: constant)
 	return Expression(lhs: lhs, relation: .GreaterThanOrEqual, rhs: rhs)
 }
